refactor(frontend): extract Cultures page dialogs into presentational components
Move the two remaining inline dialogs of the cultures page into
presentational components next to the existing dialog files in
src/pages/ (same convention, translator passed as prop):

- CulturesPublishConfirmDialog: the publish-to-public-library
  confirmation with privacy/terms links
- CulturesHistoryDialog: the history dialog for culture/project/global
  scopes with mobile card and desktop list layouts; typed with
  TFunction<'cultures'> because the culturesHistoryUtils helpers
  require it

State, data loading and the publish/restore handlers stay in
Cultures.tsx (942 -> 700 lines). No behavior change.
